\name{boot.means}
\docType{data}
\alias{boot.means}
\title{Example dataframe of mean Fst and heterozygosity from the bootstrapping for the bins.}
\description{
   Example dataframe of mean Fst and Ht from bootstrapping.
   This file contains a dataframe with 22 columns and 5 rows. 
   The columns are mean Ht, mean Fst, the number of loci in the bin, and the lower and upper bounds for each bin.
}
\usage{boot.means}
\format{data.frame}
\source{
	boot.means<-fst.boot.means(boot.out)
}
\references{
   See Flanagan & Jones
}
