\name{cis.list}
\docType{data}
\alias{cis.list}
\title{Example list of CI matrices from bootstrap output from numerical simulations}
\description{
	Example list of CI data.frames from fsthet output from numerical simulations
	The data were generated using a numerical analysis with Nm = 10, 75 demes, and 5 population samples taken. No selection was imposed.
     This is a list of data.frames containing values from 99 percent and 95 percent smoothed quantiles.
}
\usage{cis.list}
\format{list}
\source{From multiple smoothed quantile alpha thresholds, using the dataframe gpop.}
\references{
   See Flanagan & Jones
}