\name{cis}
\docType{data}
\alias{cis}
\title{Example dataframe of smoothed quantiles from fsthet}
\description{
	Example list of data.frames with smoothed quantiles from fsthet output from numerical simulations
	The data were generated using a numerical analysis with Nm = 10, 75 demes, and 5 population samples taken. No selection was imposed.
     This is a list with a single data.frame containing values from 95 percent smoothed quantiles.
}
\usage{cis}
\format{list}
\source{Ninety-five percent smoothed quantiles, using the dataframe gpop.}
\references{
   See Flanagan & Jones
}