\name{quant.out}
\docType{data}
\alias{quant.out}
\title{Example fsthet output based on numerical simulations}
\description{
   Example fsthet output based on numerical simulations
   Allelic information was generated by using a numerical analysis with Nm = 10, 75 demes, and 5 population samples taken. No selection was imposed.
   This is a list of three structures. The first is a data.frame containing the Ht and Fst values. The second is a data.frame of the bins with the lower heterozygosity values and the upper heterozygosity values for each bin. The third is a list of data.frames with the lower (Low) and upper (Upp) Fst values for each bin (the bins are in "LowHet" and "UppHet" columns.)
}
\usage{quant.out}
\format{list}
\source{Smoothed quantiles generated from the dataframe gpop.}
\references{
   See Flanagan & Jones
}