\name{remove.spaces}
\alias{remove.spaces}
\title{This removes spaces from a character vector}
\description{
  This removes spaces from a character vector. 
  It was adapted from a similar functionin adegenet.
}
\usage{
remove.spaces(charvec)
}
\arguments{
  \item{charvec}{is a vector of characters containing spaces to be removed.}
}
\value{
	\item{charvec}{A vector of characters without spaces}
}
\references{
	\url{ http://adegenet.r-forge.r-project.org/ }
}